backend/requirements.txt
flask
flask_sqlalchemy
mysql-connector-python
flask-cors